Changed Character to Unit and Unit to Friendly
