+++ Assets/Prefabs/Friendly1.prefab.meta
