+++ Assets/Prefabs/Friendly2.prefab.meta
