+++ Assets/Scripts/Friendly.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c9d52488a48f049309e274fe98c2755a
+guid: ef11185ddd7df47d089e1b77ac337fdb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
